- create a script to patch the Makefile.in and configure when using the program with AIX - correct some minor problem that appears when compiling with xlC - DiagHam has been tested on AIX version 5 with xlC version 6
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@24 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ src/HilbertSpace/QHEHilbertSpace/TrappedBosons.cc
@@ -406,7 +406,7 @@ int TrappedBosons::EvaluateHilbertSpaceDimension(int nbrBosons, int lzMax, int t
 void TrappedBosons::ErasthothenesSlieve(int maxNumber, int nbrWantedPrime, int factor)
 {
    maxNumber &= ~((int) 0x1);
-  int Size = (int) sqrt(maxNumber);
+  int Size = (int) sqrt((double) maxNumber);
   int NbrPrime = 1;
   int* PrimeNumbers = new int [Size];
   int NbrKeyFactor = 1;
